+++ app/controllers/claims_controller.rb
@@ -4,11 +4,17 @@ class ClaimsController < ApplicationController
 
   after_action :clear_claim_session, if: :submission_complete?
 
-  def create
-    claim = Claim.create!(eligibility: StudentLoans::Eligibility.new)
-    session[:claim_id] = claim.to_param
+  def new
+    render claim_page_template
+  end
 
-    redirect_to claim_path("qts-year")
+  def create
+    if update_current_claim!
+      session[:claim_id] = @current_claim.to_param
+      redirect_to claim_path(next_slug)
+    else
+      render claim_page_template
+    end
   end
 
   def show