+++ spec/features/school_search_spec.rb
@@ -3,7 +3,6 @@ require "rails_helper"
 RSpec.feature "Searching for school during Teacher Student Loan Repayments claims" do
   scenario "doesn't select a school from the search results the first time around" do
     claim = start_claim
-    choose_qts_year
 
     fill_in :school_search, with: "Penistone"
     click_on "Search"
@@ -22,7 +21,6 @@ RSpec.feature "Searching for school during Teacher Student Loan Repayments claim
 
   scenario "searches again to find school" do
     claim = start_claim
-    choose_qts_year
 
     fill_in :school_search, with: "hamp"
     click_on "Search"
@@ -41,7 +39,6 @@ RSpec.feature "Searching for school during Teacher Student Loan Repayments claim
 
   scenario "Claim school search with autocomplete", js: true do
     start_claim
-    choose_qts_year
 
     expect(page).to have_text(I18n.t("student_loans.questions.claim_school"))
     expect(page).to have_button("Search")
@@ -58,7 +55,7 @@ RSpec.feature "Searching for school during Teacher Student Loan Repayments claim
 
   scenario "Current school search with autocomplete", js: true do
     start_claim
-    choose_qts_year
+
     choose_school schools(:penistone_grammar_school)
     choose_still_teaching "Yes, at another school"
 
@@ -77,7 +74,6 @@ RSpec.feature "Searching for school during Teacher Student Loan Repayments claim
 
   scenario "School search form still works like a normal form if submitted", js: true do
     start_claim
-    choose_qts_year
 
     expect(page).to have_text(I18n.t("student_loans.questions.claim_school"))
     expect(page).to have_button("Search")
@@ -95,7 +91,6 @@ RSpec.feature "Searching for school during Teacher Student Loan Repayments claim
 
   scenario "Editing school search after autocompletion clears last selection", js: true do
     start_claim
-    choose_qts_year
 
     expect(page).to have_text(I18n.t("student_loans.questions.claim_school"))
     expect(page).to have_button("Search")
@@ -118,7 +113,6 @@ RSpec.feature "Searching for school during Teacher Student Loan Repayments claim
 
   scenario "Claim school search includes closed schools" do
     start_claim
-    choose_qts_year
 
     expect(page).to have_text(I18n.t("student_loans.questions.claim_school"))
     expect(page).to have_button("Search")
@@ -131,7 +125,6 @@ RSpec.feature "Searching for school during Teacher Student Loan Repayments claim
 
   scenario "Current school search excludes closed schools" do
     start_claim
-    choose_qts_year
     choose_school schools(:penistone_grammar_school)
     choose_still_teaching "Yes, at another school"
 
@@ -146,7 +139,6 @@ RSpec.feature "Searching for school during Teacher Student Loan Repayments claim
 
   scenario "Claim school search with autocomplete includes closed schools", js: true do
     start_claim
-    choose_qts_year
 
     expect(page).to have_text(I18n.t("student_loans.questions.claim_school"))
     expect(page).to have_button("Search")
@@ -157,7 +149,7 @@ RSpec.feature "Searching for school during Teacher Student Loan Repayments claim
 
   scenario "Current school search with autocomplete excludes closed schools", js: true do
     start_claim
-    choose_qts_year
+
     choose_school schools(:penistone_grammar_school)
     choose_still_teaching "Yes, at another school"